Implement authentication routes.
Here's an explanation:

The `<SpotifyAuthProvider />` component provides to all of its child
components a token for the upcoming requests.

A `<PrivateRoute />` component is used to redirect to the "/login" route
if there is not a token in localStorage.

At the "/login" URL, the <LoginPage /> componenet is rendered, which has
a button to a spotify endpoint responsible for the authentication.

After the user grants or declines access, Spoitfy will redirect to our
app at the "/auth" URL with the access_token in it (as well as the
expiration time etc.).

We just handle how to get the access token from the URL (which is done
inside the `<AuthPage />` component, see function `getToken`) and change
the `<SpotifyAuthProvider />` token from it.

After that, the user will be authenticated and it will be redirected to
its home page ("/").