Remove SpotifyListItem, trying to make things less abstract
It's no good to use such an overloaded component because if I want to
add a feature for a particular case (a track, for example), it would be
become too patchy.

Instead, compose the lists bare Material-UI components, and use a
component SpotifyListSkeleton to handle when data is being loaded.